Issue #90: DST causes 404: Testcases
Fixed testcases that became broken due to a change in the json-response.
Previously, recurring event dates were returned in local time, whereas non-recurring event dates were returned in UTC. This was made consistent but the testcases were not updated to reflect this.

Moved the has_occurence_at methods as per suggestion. Also added its missing testcase.
